+++ src/main.tsx
@@ -4,6 +4,7 @@ import { createRoot } from 'react-dom/client';
 // import App from './App';
 // import { store } from './app/store';
 import './index.css';
+import HookedTextField from './features/textField/HookedTextField';
 
 const container = document.getElementById('root');
 
@@ -13,7 +14,10 @@ if (container) {
   root.render(
     <React.StrictMode>
       {/* <Provider store={store}> */}
-      <h1>Main</h1>
+      <main>
+        <HookedTextField type="float" name="deliveryDistance" />
+      </main>
+
       {/* </Provider> */}
     </React.StrictMode>,
   );